Added a dependency range instead of using LASTEST
* Switched dependencies to use an xml file as well
* Renamed editor script to include Android in the name
