+++ OneSignalExample/Assets/OneSignal/Editor/OneSignalDependencies.xml
@@ -1,0 +1,8 @@
+<dependencies>
+  <androidPackages>
+    <androidPackage spec="com.google.android.gms:play-services-gcm:[10.2.1, 11.9.0)" />
+    <androidPackage spec="com.google.android.gms:play-services-location:[10.2.1, 11.9.0)" />
+    <androidPackage spec="com.android.support:support-v4:[26.0.0, 27.2.0)" />
+    <androidPackage spec="com.android.support:customtabs:[26.0.0, 27.2.0)" />
+  </androidPackages>
+</dependencies>
