+++ OneSignalExample/Assets/OneSignal/Editor/OneSignalDependencies.xml.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 127a9573f8846469b8feaf4dd86f9587
+timeCreated: 1521620682
+licenseType: Free
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
